requirements.txt
certifi==2019.11.28
prometheus-client==0.6.0
requests==2.22.0
pyyaml==5.3
schema==0.7.1
apscheduler==3.6.3